move generation from variant class
