+++ modules/draughts/src/main/variant/Antidraughts.scala
@@ -13,6 +13,8 @@ case object Antidraughts extends Variant(
 ) {
 
   def pieces = Standard.pieces
+  def captureDirs = Standard.captureDirs
+  def moveDirsColor = Standard.moveDirsColor
 
   // Only difference is that you win when you run out of moves (no pieces or all blocked)
   override def winner(situation: Situation): Option[Color] =
