+++ modules/draughts/src/main/variant/Standard.scala
@@ -13,5 +13,6 @@ case object Standard extends Variant(
 ) {
 
   val pieces: Map[Pos, Piece] = Variant.symmetricFourRank(Vector(Man, Man, Man, Man, Man), boardSize)
-
+  val captureDirs: Directions = List((Variant.UpLeft, _.moveUpLeft), (Variant.UpRight, _.moveUpRight), (Variant.DownLeft, _.moveDownLeft), (Variant.DownRight, _.moveDownRight))
+  val moveDirsColor: Map[Color, Directions] = Map(White -> List((Variant.UpLeft, _.moveUpLeft), (Variant.UpRight, _.moveUpRight)), Black -> List((Variant.DownLeft, _.moveDownLeft), (Variant.DownRight, _.moveDownRight)))
 }
