+++ modules/draughts/src/main/variant/Variant.scala
@@ -1,9 +1,9 @@
 package draughts
 package variant
 
+import scala.annotation.tailrec
 import scala.collection.breakOut
 import scalaz.Validation.FlatMap._
-import scalaz.Validation.failureNel
 
 // Correctness depends on singletons for each variant ID
 abstract class Variant private[variant] (
@@ -18,6 +18,8 @@ abstract class Variant private[variant] (
 ) {
 
   def pieces: Map[Pos, Piece]
+  def captureDirs: Directions
+  def moveDirsColor: Map[Color, Directions]
 
   def standard = this == Standard
   def frisian = this == Frisian
@@ -28,7 +30,6 @@ abstract class Variant private[variant] (
   def fromPosition = this == FromPosition
 
   def frisianVariant = frisian || frysk
-
   def exotic = !standard
 
   def initialFen = format.Forsyth.initial
@@ -39,16 +40,14 @@ abstract class Variant private[variant] (
     case _ => false
   }
 
+  def getCaptureValue(board: Board, taken: List[Pos]) = taken.length
+  def getCaptureValue(board: Board, taken: Pos) = 1
+
   def allCaptures(situation: Situation): Map[Pos, List[Move]] =
     situation.actors.collect {
       case actor if actor.captures.nonEmpty =>
         actor.pos -> actor.captures
     }(breakOut)
-
-  @inline
-  def captureValue(board: Board, taken: List[Pos]) = taken.length
-  @inline
-  def captureValue(board: Board, taken: Pos) = 1
 
   def validMoves(situation: Situation, finalSquare: Boolean = false): Map[Pos, List[Move]] = {
 
@@ -57,7 +56,7 @@ abstract class Variant private[variant] (
     for (actor <- situation.actors) {
       val capts = if (finalSquare) actor.capturesFinal else actor.captures
       if (capts.nonEmpty) {
-        val lineValue = capts.head.taken.fold(0)(captureValue(situation.board, _))
+        val lineValue = capts.head.taken.fold(0)(_.length)
         if (lineValue > bestLineValue) {
           bestLineValue = lineValue
           captureMap = Map(actor.pos -> capts)
@@ -95,10 +94,6 @@ abstract class Variant private[variant] (
       actor.noncaptures
   }
 
-  /*def longRangeThreatens(board: Board, p: Pos, dir: Direction, to: Pos): Boolean = dir._2(p) exists { next =>
-    next == to || (!board.pieces.contains(next) && longRangeThreatens(board, next, dir, to))
-  }*/
-
   def move(situation: Situation, from: Pos, to: Pos, promotion: Option[PromotableRole], finalSquare: Boolean = false, forbiddenUci: Option[List[String]] = None, captures: Option[List[Pos]] = None, partialCaptures: Boolean = false): Valid[Move] = {
 
     // Find the move in the variant specific list of valid moves
@@ -112,12 +107,156 @@ abstract class Variant private[variant] (
 
     for {
       actor ← situation.board.actors get from toValid "No piece on " + from
-      myActor ← actor.validIf(actor is situation.color, "Not my piece on " + from)
+      _ ← actor.validIf(actor is situation.color, "Not my piece on " + from)
       m1 ← findMove(from, to) toValid "Piece on " + from + " cannot move to " + to
       m2 ← m1 withPromotion promotion toValid "Piece on " + from + " cannot promote to " + promotion
       m3 <- m2 validIf (isValidPromotion(promotion), "Cannot promote to " + promotion + " in this game mode")
     } yield m3
 
+  }
+
+  def shortRangeCaptures(actor: Actor, finalSquare: Boolean): List[Move] = {
+    val buf = new scala.collection.mutable.ArrayBuffer[Move]
+    var bestCaptureValue = 0
+
+    def walkCaptures(walkDir: Direction, curBoard: Board, curPos: PosMotion, destPos: Option[PosMotion], destBoard: Option[Board], allSquares: List[Pos], allTaken: List[Pos], captureValue: Int): Unit =
+      walkDir._2(curPos).fold() {
+        nextPos =>
+          curBoard(nextPos) match {
+            case Some(captPiece) if captPiece.isNot(actor.color) && !captPiece.isGhost =>
+              walkDir._2(nextPos) match {
+                case Some(landingPos) if curBoard(landingPos).isEmpty =>
+                  curBoard.taking(curPos, landingPos, nextPos).fold() { boardAfter =>
+                    val newSquares = landingPos :: allSquares
+                    val newTaken = nextPos :: allTaken
+                    val newCaptureValue = captureValue + getCaptureValue(actor.board, nextPos)
+                    if (newCaptureValue > bestCaptureValue) {
+                      bestCaptureValue = newCaptureValue
+                      buf.clear()
+                    }
+                    if (newCaptureValue == bestCaptureValue) {
+                      if (finalSquare)
+                        buf += actor.move(landingPos, boardAfter.withoutGhosts, newSquares, newTaken)
+                      else
+                        buf += actor.move(destPos.getOrElse(landingPos), destBoard.getOrElse(boardAfter), newSquares, newTaken)
+                    }
+                    val opposite = Variant.oppositeDirs(walkDir._1)
+                    captureDirs.foreach {
+                      captDir =>
+                        if (captDir._1 != opposite)
+                          walkCaptures(captDir, boardAfter, landingPos, destPos.getOrElse(landingPos).some, destBoard.getOrElse(boardAfter).some, newSquares, newTaken, newCaptureValue)
+                    }
+                  }
+                case _ =>
+              }
+            case _ =>
+          }
+      }
+
+    captureDirs.foreach {
+      walkCaptures(_, actor.board, actor.pos, None, None, Nil, Nil, 0)
+    }
+
+    buf.flatMap { m =>
+      if (finalSquare || m.capture.exists(_.length == 1)) actor.maybePromote(m)
+      else m.some
+    } toList
+  }
+
+  def longRangeCaptures(actor: Actor, finalSquare: Boolean): List[Move] = {
+    val buf = new scala.collection.mutable.ArrayBuffer[Move]
+    var bestCaptureValue = 0
+
+    // "transposition table", dramatically reduces calculation time for extreme frisian positions like W:WK50:B3,7,10,12,13,14,17,20,21,23,25,30,32,36,38,39,41,43,K47
+    val cacheExtraCapts = scala.collection.mutable.LongMap.empty[Int]
+    // but not enough apparently, frisian can still max out CPU with e.g. W:WK5:BK2,K4,K7,K8,K9,K10,K11,K13,K15,K16,K18,K19,K20,K21,K22,K24,K27,K29,K30,K31,K32,K33,K35,K36,K38,K40,K41,K42,K43,K44,K47,K49
+    // temporary hackfix for server stability: simply abort on * wet finger * 1600 cache entries - this should be enough for any practical game position, but may lead to incorrect dests in extreme frisian analysis (soit)
+    val maxCache = 1600
+
+    @tailrec
+    def walkUntilCapture(walkDir: Direction, curBoard: Board, curPos: PosMotion, destPos: Option[PosMotion], destBoard: Option[Board], allSquares: List[Pos], allTaken: List[Pos], captureValue: Int): Int =
+      if (cacheExtraCapts.size > maxCache) 0
+      else walkDir._2(curPos) match {
+        case Some(nextPos) =>
+          curBoard(nextPos) match {
+            case None =>
+              curBoard.move(curPos, nextPos) match {
+                case Some(boardAfter) =>
+                  walkUntilCapture(walkDir, boardAfter, nextPos, destPos, destBoard, allSquares, allTaken, captureValue)
+                case _ =>
+                  captureValue
+              }
+            case Some(captPiece) if captPiece.isNot(actor.color) && !captPiece.isGhost =>
+              walkDir._2(nextPos) match {
+                case Some(landingPos) if curBoard(landingPos).isEmpty =>
+                  curBoard.taking(curPos, landingPos, nextPos) match {
+                    case Some(boardAfter) =>
+                      walkAfterCapture(walkDir, boardAfter, landingPos, destPos, destBoard, allSquares, nextPos :: allTaken, captureValue + getCaptureValue(actor.board, nextPos))
+                    case _ =>
+                      captureValue
+                  }
+                case _ => captureValue
+              }
+            case _ => captureValue
+          }
+        case _ => captureValue
+      }
+
+    def walkAfterCapture(walkDir: Direction, curBoard: Board, curPos: PosMotion, destPos: Option[PosMotion], destBoard: Option[Board], allSquares: List[Pos], newTaken: List[Pos], newCaptureValue: Int): Int = {
+      val captsHash = curBoard.pieces.hashCode() + walkDir._1
+      val cachedExtraCapts = cacheExtraCapts.get(captsHash)
+      if (cacheExtraCapts.size > maxCache) 0
+      else cachedExtraCapts match {
+        case Some(extraCapts) if newCaptureValue + extraCapts < bestCaptureValue =>
+          // no need to calculate lines where we know they will end up too short
+          newCaptureValue + extraCapts
+        case _ =>
+          val newSquares = curPos :: allSquares
+          if (newCaptureValue > bestCaptureValue) {
+            bestCaptureValue = newCaptureValue
+            buf.clear()
+          }
+          if (newCaptureValue == bestCaptureValue) {
+            if (finalSquare)
+              buf += actor.move(curPos, curBoard.withoutGhosts, newSquares, newTaken)
+            else
+              buf += actor.move(destPos.getOrElse(curPos), destBoard.getOrElse(curBoard), newSquares, newTaken)
+          }
+          val opposite = Variant.oppositeDirs(walkDir._1)
+          var maxExtraCapts = 0
+          captureDirs.foreach { captDir =>
+            if (captDir._1 != opposite) {
+              val extraCapture = walkUntilCapture(captDir, curBoard, curPos, destPos.getOrElse(curPos).some, destBoard.getOrElse(curBoard).some, newSquares, newTaken, newCaptureValue) - newCaptureValue
+              if (extraCapture > maxExtraCapts)
+                maxExtraCapts = extraCapture
+            }
+          }
+          walkDir._2(curPos) match {
+            case Some(nextPos) =>
+              curBoard.move(curPos, nextPos) match {
+                case Some(boardAfter) =>
+                  val extraCapture = walkAfterCapture(walkDir, boardAfter, nextPos, destPos, destBoard, allSquares, newTaken, newCaptureValue) - newCaptureValue
+                  if (extraCapture > maxExtraCapts)
+                    maxExtraCapts = extraCapture
+                case _ =>
+              }
+            case _ =>
+          }
+          if (cachedExtraCapts.isEmpty)
+            cacheExtraCapts += (captsHash, maxExtraCapts)
+          newCaptureValue + maxExtraCapts
+      }
+    }
+
+    captureDirs.foreach {
+      walkUntilCapture(_, actor.board, actor.pos, None, None, Nil, Nil, 0)
+    }
+
+    if (cacheExtraCapts.size > maxCache) {
+      logger.warn(s"longRangeCaptures aborted with ${cacheExtraCapts.size} entries for ${actor.piece} at ${actor.pos.shortKey} on ${draughts.format.Forsyth.exportBoard(actor.board)}")
+    }
+
+    buf.toList
   }
 
   def checkmate(situation: Situation) = situation.validMoves.isEmpty
@@ -194,13 +333,8 @@ abstract class Variant private[variant] (
   def valid(board: Board, strict: Boolean) = Color.all forall validSide(board, strict)_
 
   val roles = List(Man, King)
-
-  val promotableRoles: List[PromotableRole] = List(King)
-
   lazy val rolesByPdn: Map[Char, Role] = roles.map { r => (r.pdn, r) }(breakOut)
 
-  val captureDirs: Directions = List((Actor.UpLeft, _.moveUpLeft), (Actor.UpRight, _.moveUpRight), (Actor.DownLeft, _.moveDownLeft), (Actor.DownRight, _.moveDownRight))
-  val moveDirsColor: Map[Color, Directions] = Map(White -> List((Actor.UpLeft, _.moveUpLeft), (Actor.UpRight, _.moveUpRight)), Black -> List((Actor.DownLeft, _.moveDownLeft), (Actor.DownRight, _.moveDownRight)))
   val moveDirsAll: Directions = moveDirsColor(White) ::: moveDirsColor(Black)
 
   override def toString = s"Variant($name)"
@@ -245,6 +379,27 @@ object Variant {
     draughts.variant.Breakthrough,
     draughts.variant.Russian,
     draughts.variant.FromPosition
+  )
+
+  val UpLeft = 1
+  val UpRight = 2
+  val DownLeft = 3
+  val DownRight = 4
+  val Up = 5
+  val Down = 6
+  val Left = 7
+  val Right = 8
+
+  val oppositeDirs: Array[Int] = Array(
+    0,
+    DownRight,
+    DownLeft,
+    UpRight,
+    UpLeft,
+    Down,
+    Up,
+    Right,
+    Left
   )
 
   private[variant] def symmetricFourRank(rank: IndexedSeq[Role], boardSize: Board.BoardSize): Map[Pos, Piece] = {
